MAINT: add convex_analysis function from xsf
